Moved classes from request.runner package back into the request package due to access in CachedSpiceRequest
